Move component PrinterButton from 'components' to 'components/flow'
